fixed some errrors, made editor resolution not weird, and made the mask texture work
