+++ Assets/AccessingTheUnityFileExplorer-main/FileExplorerUnityPackage.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: efbe6767b8675a947a4ff30ca1aaa18a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
