+++ Assets/Plugins/Unity-RectTransform-Preset-Utils-master(1)/Unity-RectTransform-Preset-Utils-master/AnchorPresetUtilsTestProject.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6cf37fbf7d630de4fbc5af6b55927162
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
