+++ Assets/Prefabs/canvases/canvasMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d39d39783b792504f86fe80cab5dcfe4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
