+++ Assets/Prefabs/canvases/schematicMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df90d28e2265546ab8ed8b497314112d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
