--- Assets/Prefabs/schematics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 708f44bf30b3344dbbeff0f6affeb861
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
