--- Assets/Prefabs/schematics/maskMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 173fc66cf05114d4fb081aa6d29b47dd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
