+++ Assets/Unity-RectTransform-Preset-Utils-master/Unity-RectTransform-Preset-Utils-master/AnchorPresetUtilsTestProject.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dbaa2f19038df284882f47261a7d8840
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
