--- Assets/schematicManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dafd776d793ed4aa58a26e6ffcda5392
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
