+++ UserSettings/EditorUserSettings.asset
@@ -15,19 +15,19 @@ EditorUserSettings:
       value: 0657555156505c5e5f5b547742725c441415487d7c7177602e7d1b35b5e3653b
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 010751555d030a5e080d587047710c4444161b79287a7e66287d4a64e0e1663a
-      flags: 0
-    RecentlyUsedSceneGuid-4:
       value: 51050d5703575103580f5a2040260d44434e4c722d7a76627b7f4460b3e13568
       flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-4:
       value: 550102555056500b55565b7011215c4441161d78287825362e2c1c62e1b5616b
       flags: 0
-    RecentlyUsedSceneGuid-6:
+    RecentlyUsedSceneGuid-5:
       value: 025357550651590b5f5e5b7a44210b4413154c2e757022347f7a1c35b2b6303a
       flags: 0
-    RecentlyUsedSceneGuid-7:
+    RecentlyUsedSceneGuid-6:
       value: 02550c5204545a080c5809754672074412164b282a702064747b4b61b4b3616b
+      flags: 0
+    RecentlyUsedSceneGuid-7:
+      value: 010751555d030a5e080d587047710c4444161b79287a7e66287d4a64e0e1663a
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 51555500540d08085c5b5c7b42720944404f4f7f2e2b2460752a1961bab6646e
